refactor(ui): storybook관리 및 utils분리
